+++ src/lib/ffi/ffi_kdf.cpp
@@ -76,7 +76,7 @@ int botan_bcrypt_generate(uint8_t* out, size_t* out_len,
          throw FFI_Error("Bad bcrypt work factor " + std::to_string(wf));
 
       Botan::RandomNumberGenerator& rng = safe_get(rng_obj);
-      const std::string bcrypt = Botan::generate_bcrypt(pass, rng, wf);
+      const std::string bcrypt = Botan::generate_bcrypt(pass, rng, static_cast<uint16_t>(wf));
       return write_str_output(out, out_len, bcrypt);
       });
 #else
